LPS-108310 portal-search-elasticsearch6-impl: extract logic into IndexSettingsContributorHelper, prepare for next
